server-side/requirements.txt
flask
python-dotenv
openai
langchain
langchain-openai
langchain-google-genai
pandas
numpy
pymongo
flask-sqlalchemy
flask-bcrypt
flask-session
flask-cors
flask-migrate
deep-translator
langdetect
weasyprint
tavily-python
lingua-language-detector
iso639-lang
reportlab
sentence-transformers
gtts
pytz
google-search-results
faiss-cpu
pypdf
google-generativeai
langchain-community
scrapfly-sdk
PyMuPDF
Pillow